#119: Move `smoketests` out of `src`
* Move the smoketests folder out of src to match the tests folder.
