+++ .gitignore
@@ -15,7 +15,7 @@ dist-ssr
 .vite/vitest/
 
 # ignore smoketest log
-.src/smoketest/testLog.txt
+my-webxr-app/smoketests/testLog.txt
 
 ### https://github.com/github/gitignore/blob/main/Node.gitignore
 # Logs
