+++ my-webxr-app/package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "scripts": {
     "dev": "vite",
-    "test": "vitest run --exclude 'src/smoketest/**'",
+    "test": "vitest run --exclude 'smoketests/**'",
     "smoketest": "vitest run",
     "test:watch": "vitest watch",
     "coverage": "vitest run --coverage",
