+++ my-webxr-app/smoketests/DisplayPointsSmokeTest.test.tsx
@@ -1,4 +1,3 @@
-// eslint-disable-next-line check-file/filename-naming-convention
 import ContainText from './TextInFileSearch';
 
 describe('Data Point Component in WebXR Application', () => {
