+++ my-webxr-app/smoketests/ImportingFromCSV.test.tsx
@@ -1,4 +1,3 @@
-// eslint-disable-next-line check-file/filename-naming-convention
 import ContainText from './TextInFileSearch';
 
 describe('Importing to the web app from a CSV, to then use the data. This can come from an exterior url or a file import.', () => {
