+++ my-webxr-app/smoketests/TestHookWrite.tsx
