+++ my-webxr-app/smoketests/TextInFileSearchtest.tsx
