+++ my-webxr-app/smoketests/testLog.txt.template
@@ -1,0 +1,1 @@
+# For smoketests, rename this file to `testLog.txt` and add contents from the test to be checked
