+++ my-webxr-app/tsconfig.json
@@ -26,7 +26,8 @@
   "include": [
     ".eslintrc.cjs",
     "src",
-    "tests"
+    "tests",
+    "smoketests"
   ],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
